separate destination lookup pools
git-svn-id: http://svn.vuze.com/public/plugins@40665 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ networks/azi2phelper/org/parg/azureus/plugins/networks/i2p/I2PHelperDHT.java
@@ -91,7 +91,7 @@ I2PHelperDHT
 	get(
 		byte[] 						ih, 
 		String						reason,
-		byte						flags,
+		short						flags,
 		int 						max, 
 		long						maxWait,
 		I2PHelperDHTListener		listener );
@@ -100,7 +100,7 @@ I2PHelperDHT
 	put(
 		byte[] 						ih, 
 		String						reason,
-		byte						flags,
+		short						flags,
 		I2PHelperDHTListener		listener );
 	
 	public void
